# api/openapi.yaml
openapi: 3.1.0
info:
  title: Semantic Dropdown Search API
  description: Deterministic semantic dropdown search, with optional embedding-based
    enhancement.
  version: 3.0.0
paths:
  /semantic-config:
    get:
      summary: Get Config
      description: Returns the current API configuration.
      operationId: get_config_semantic_config_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIConfig'
    patch:
      summary: Update Config
      description: Updates API configuration fields only after dependencies are ready.
      operationId: update_config_semantic_config_patch
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SemanticConfigUpdate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIConfig'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /semantic-index:
    post:
      summary: Index Items
      description: Indexes a batch of items. Computes embeddings if enabled.
      operationId: index_items_semantic_index_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IndexRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /semantic-search:
    get:
      summary: Search
      description: 'Searches indexed content.

        Returns ranked results based on the chosen mode (deterministic, embedding,
        hybrid).'
      operationId: search_semantic_search_get
      parameters:
      - name: q
        in: query
        required: true
        schema:
          type: string
          description: The plain-text query to search for
          title: Q
        description: The plain-text query to search for
      - name: mode
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Override engine_mode for this request
          title: Mode
        description: Override engine_mode for this request
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    APIConfig:
      properties:
        engine_mode:
          type: string
          enum:
          - deterministic
          - embedding
          - hybrid
          title: Engine Mode
          description: 'Mode of the search engine: deterministic, embedding, or hybrid.'
          default: deterministic
        embedding_enabled:
          type: boolean
          title: Embedding Enabled
          description: Whether to use the embedding model.
          default: false
        embedding_model:
          type: string
          title: Embedding Model
          description: Name or path of the embedding model to use.
          default: sentence-transformers/all-MiniLM-L6-v2
        max_results:
          type: integer
          minimum: 0.0
          title: Max Results
          description: Maximum number of results to return per query.
          default: 10
        fallback_keyword_search:
          type: boolean
          title: Fallback Keyword Search
          description: Whether to fallback to keyword search if other methods fail
            or have low scores.
          default: true
        storage_backend:
          type: string
          enum:
          - memory
          - sqlite
          title: Storage Backend
          description: Which storage backend to use for indexing and retrieval.
          default: memory
      type: object
      title: APIConfig
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    IndexRequest:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ItemInput'
          type: array
          title: Items
      type: object
      required:
      - items
      title: IndexRequest
    ItemInput:
      properties:
        id:
          anyOf:
          - type: string
          - type: 'null'
          title: Id
        text:
          type: string
          title: Text
        tags:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Tags
        descriptor:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Descriptor
      type: object
      required:
      - text
      title: ItemInput
    SearchResponse:
      properties:
        query:
          type: string
          title: Query
        results:
          items:
            $ref: '#/components/schemas/SearchResult'
          type: array
          title: Results
        mode:
          type: string
          title: Mode
      type: object
      required:
      - query
      - results
      - mode
      title: SearchResponse
    SearchResult:
      properties:
        id:
          type: string
          title: Id
        text:
          type: string
          title: Text
        score:
          type: number
          title: Score
        descriptor:
          additionalProperties: true
          type: object
          title: Descriptor
      type: object
      required:
      - id
      - text
      - score
      - descriptor
      title: SearchResult
    SemanticConfigUpdate:
      properties:
        engine_mode:
          anyOf:
          - type: string
          - type: 'null'
          title: Engine Mode
        embedding_enabled:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Embedding Enabled
        embedding_model:
          anyOf:
          - type: string
          - type: 'null'
          title: Embedding Model
        max_results:
          anyOf:
          - type: integer
            minimum: 0.0
          - type: 'null'
          title: Max Results
        fallback_keyword_search:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Fallback Keyword Search
      type: object
      title: SemanticConfigUpdate
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError